refactor(state): segment useAppState into 4 domain hooks (v3.3.0)
- useUIState: modals, panels, tabs, markup
- useSongMeta: editorial metadata
- useMusicalMeta: musical parameters
- useSessionState: runtime, persistence, API key
- useAppState: barrel re-export (zero breaking change)
